requirements.txt
paramiko==3.4.0
python-dotenv==1.0.1
requests==2.31.0
python-telegram-bot==13.7
psycopg2==2.9.9